discover: pxe: Avoid dereferencing null pointer
When result is null, we may end up in the error handling path where we
try to dereference null to call cleanup_local. This adds a check for
result.

Found with scan-build.

	[Fixed up commit message typo]

+++ discover/pxe-parser.c
@@ -324,7 +324,7 @@ static void pxe_conf_parse_cb(struct load_url_result *result, void *data)
 
 	talloc_free(buf);
 out_clean:
-	if (result->cleanup_local)
+	if (result && result->cleanup_local)
 		unlink(result->local);
 	talloc_free(conf);
 }
